refactor: refactor graphic-service
