+++ packages/vrender-core/src/graphic/arc.ts
@@ -1,13 +1,14 @@
-import type { AABBBounds } from '@visactor/vutils';
+import type { AABBBounds, IAABBBounds } from '@visactor/vutils';
 import { pi2, sin, epsilon, abs, asin, clampAngleByRadian, isNumber, cos, sqrt, isArray } from '@visactor/vutils';
 import type { IArc, IArcGraphicAttribute } from '../interface/graphic/arc';
 import { Graphic, GRAPHIC_UPDATE_TAG_KEY, NOWORK_ANIMATE_ATTR } from './graphic';
 import { CustomPath2D } from '../common/custom-path2d';
-import { parsePadding } from '../common/utils';
+import { circleBounds, parsePadding } from '../common/utils';
 import { getTheme } from './theme';
 import { application } from '../application';
 import type { GraphicType } from '../interface';
 import { ARC_NUMBER_TYPE } from './constants';
+import { updateBoundsOfCommonOuterBorder } from './graphic-service/common-outer-boder-bounds';
 
 /**
  * 部分代码参考 https://github.com/d3/d3-shape/
@@ -107,7 +108,7 @@ export class Arc extends Graphic<IArcGraphicAttribute> implements IArc {
   }
 
   getParsedCornerRadius() {
-    const arcTheme = getTheme(this).arc;
+    const arcTheme = this.getGraphicTheme();
     const {
       cornerRadius = arcTheme.cornerRadius,
       innerPadding = arcTheme.innerPadding,
@@ -143,7 +144,7 @@ export class Arc extends Graphic<IArcGraphicAttribute> implements IArc {
   }
 
   getParsedAngle() {
-    const arcTheme = getTheme(this).arc;
+    const arcTheme = this.getGraphicTheme();
     let { startAngle = arcTheme.startAngle, endAngle = arcTheme.endAngle } = this.attribute;
     const { cap = arcTheme.cap } = this.attribute;
 
@@ -157,8 +158,8 @@ export class Arc extends Graphic<IArcGraphicAttribute> implements IArc {
       let startCap = 1;
       let endCap = 1;
       if ((cap as boolean[]).length) {
-        startCap = Number(cap[0]);
-        endCap = Number(cap[1]);
+        startCap = Number((cap as boolean[])[0]);
+        endCap = Number((cap as boolean[])[1]);
       }
       let { outerRadius = arcTheme.outerRadius, innerRadius = arcTheme.innerRadius } = this.attribute;
       const { outerPadding = arcTheme.outerPadding, innerPadding = arcTheme.innerPadding } = this.attribute;
@@ -186,7 +187,7 @@ export class Arc extends Graphic<IArcGraphicAttribute> implements IArc {
 
   // 参考 https://github.com/d3/d3-shape/blob/main/src/arc.js
   getParsePadAngle(startAngle: number, endAngle: number) {
-    const arcTheme = getTheme(this).arc;
+    const arcTheme = this.getGraphicTheme();
     const {
       innerPadding = arcTheme.innerPadding,
       outerPadding = arcTheme.outerPadding,
@@ -251,25 +252,91 @@ export class Arc extends Graphic<IArcGraphicAttribute> implements IArc {
     };
   }
 
-  protected doUpdateAABBBounds(full?: boolean): AABBBounds {
-    const arcTheme = getTheme(this).arc;
-    this._AABBBounds.clear();
-    const attribute = this.attribute;
-    const bounds = application.graphicService.updateArcAABBBounds(
-      attribute,
-      getTheme(this).arc,
-      this._AABBBounds,
-      full,
-      this
-    ) as AABBBounds;
+  getGraphicTheme() {
+    return getTheme(this).arc;
+  }
 
-    const { boundsPadding = arcTheme.boundsPadding } = attribute;
-    const paddingArray = parsePadding(boundsPadding);
-    if (paddingArray) {
-      bounds.expand(paddingArray);
+  protected updateAABBBounds(
+    attribute: IArcGraphicAttribute,
+    arcTheme: Required<IArcGraphicAttribute>,
+    aabbBounds: IAABBBounds,
+    full?: boolean
+  ) {
+    if (!application.graphicService.validCheck(attribute, arcTheme, aabbBounds, this)) {
+      return aabbBounds;
     }
-    this.clearUpdateBoundTag();
-    return bounds;
+    if (!this.updatePathProxyAABBBounds(aabbBounds)) {
+      full
+        ? this.updateArcAABBBoundsImprecise(attribute, arcTheme, aabbBounds)
+        : this.updateArcAABBBoundsAccurate(attribute, arcTheme, aabbBounds);
+    }
+
+    const tb1 = application.graphicService.tempAABBBounds1;
+    const tb2 = application.graphicService.tempAABBBounds2;
+    tb1.setValue(aabbBounds.x1, aabbBounds.y1, aabbBounds.x2, aabbBounds.y2);
+    tb2.setValue(aabbBounds.x1, aabbBounds.y1, aabbBounds.x2, aabbBounds.y2);
+
+    updateBoundsOfCommonOuterBorder(attribute, arcTheme, tb1);
+    aabbBounds.union(tb1);
+    tb1.setValue(tb2.x1, tb2.y1, tb2.x2, tb2.y2);
+
+    const { lineJoin = arcTheme.lineJoin } = attribute;
+    application.graphicService.transformAABBBounds(attribute, aabbBounds, arcTheme, lineJoin === 'miter', this);
+
+    return aabbBounds;
+  }
+
+  protected updateArcAABBBoundsImprecise(
+    attribute: IArcGraphicAttribute,
+    arcTheme: Required<IArcGraphicAttribute>,
+    aabbBounds: IAABBBounds
+  ): IAABBBounds {
+    let { outerRadius = arcTheme.outerRadius, innerRadius = arcTheme.innerRadius } = attribute;
+    const { outerPadding = arcTheme.outerPadding, innerPadding = arcTheme.innerPadding } = attribute;
+    outerRadius += outerPadding;
+    innerRadius -= innerPadding;
+    if (outerRadius < innerRadius) {
+      outerRadius = innerRadius;
+    }
+
+    aabbBounds.set(-outerRadius, -outerRadius, outerRadius, outerRadius);
+
+    return aabbBounds;
+  }
+  protected updateArcAABBBoundsAccurate(
+    attribute: IArcGraphicAttribute,
+    arcTheme: Required<IArcGraphicAttribute>,
+    aabbBounds: IAABBBounds
+  ): IAABBBounds {
+    let { outerRadius = arcTheme.outerRadius, innerRadius = arcTheme.innerRadius } = attribute;
+    const { outerPadding = arcTheme.outerPadding, innerPadding = arcTheme.innerPadding } = attribute;
+    outerRadius += outerPadding;
+    innerRadius -= innerPadding;
+    if (outerRadius < innerRadius) {
+      // 不用解构，避免性能问题
+      const temp = outerRadius;
+      outerRadius = innerRadius;
+      innerRadius = temp;
+    }
+    let { endAngle = arcTheme.endAngle, startAngle = arcTheme.startAngle } = attribute;
+
+    if (startAngle > endAngle) {
+      const temp = startAngle;
+      startAngle = endAngle;
+      endAngle = temp;
+    }
+
+    if (outerRadius <= epsilon) {
+      aabbBounds.set(0, 0, 0, 0);
+    } else if (Math.abs(endAngle - startAngle) > pi2 - epsilon) {
+      aabbBounds.set(-outerRadius, -outerRadius, outerRadius, outerRadius);
+    } else {
+      // 直接内外两个radius叠加即可，不需要更精确
+      circleBounds(startAngle, endAngle, outerRadius, aabbBounds);
+      circleBounds(startAngle, endAngle, innerRadius, aabbBounds);
+    }
+
+    return aabbBounds;
   }
 
   protected needUpdateTags(keys: string[]): boolean {
@@ -277,11 +344,6 @@ export class Arc extends Graphic<IArcGraphicAttribute> implements IArc {
   }
   protected needUpdateTag(key: string): boolean {
     return super.needUpdateTag(key, ARC_UPDATE_TAG_KEY);
-  }
-
-  getDefaultAttribute(name: string) {
-    const arcTheme = getTheme(this).arc;
-    return arcTheme[name];
   }
 
   toCustomPath() {
