+++ packages/vrender-core/src/graphic/glyph.ts
@@ -1,4 +1,4 @@
-import type { AABBBounds, IPointLike } from '@visactor/vutils';
+import type { AABBBounds, IAABBBounds, IPointLike } from '@visactor/vutils';
 import { Graphic, NOWORK_ANIMATE_ATTR } from './graphic';
 import type {
   GraphicType,
@@ -156,14 +156,31 @@ export class Glyph extends Graphic<IGlyphGraphicAttribute> implements IGlyph {
     return this;
   }
 
+  getGraphicTheme() {
+    return getTheme(this).glyph;
+  }
+
+  protected updateAABBBounds(
+    attribute: IGlyphGraphicAttribute,
+    theme: Required<IGlyphGraphicAttribute>,
+    aabbBounds: IAABBBounds
+  ) {
+    if (!application.graphicService.validCheck(attribute, theme, aabbBounds, this)) {
+      return aabbBounds;
+    }
+    // 添加子节点
+    this.getSubGraphic().forEach((node: IGraphic) => {
+      aabbBounds.union(node.AABBBounds);
+    });
+
+    // glyph不需要计算AABBBounds
+    // this.transformAABBBounds(attribute, aabbBounds, theme, graphic);
+    return aabbBounds;
+  }
+
   protected doUpdateAABBBounds(): AABBBounds {
     this._AABBBounds.clear();
-    const bounds = application.graphicService.updateGlyphAABBBounds(
-      this.attribute,
-      getTheme(this).glyph,
-      this._AABBBounds,
-      this
-    ) as AABBBounds;
+    const bounds = this.updateAABBBounds(this.attribute, this.getGraphicTheme(), this._AABBBounds) as AABBBounds;
     this.clearUpdateBoundTag();
     return bounds;
   }
