+++ packages/vrender-core/src/graphic/graphic.ts
@@ -1,5 +1,5 @@
 import type { ICustomPath2D } from './../interface/path';
-import type { OBBBounds, Dict, IPointLike, IAABBBounds } from '@visactor/vutils';
+import type { Dict, IPointLike, IAABBBounds } from '@visactor/vutils';
 import {
   AABBBounds,
   Matrix,
@@ -42,9 +42,11 @@ import { application } from '../application';
 import { Animate, DefaultStateAnimateConfig } from '../animate';
 import { interpolateColor } from '../color-string/interpolate';
 import { CustomPath2D } from '../common/custom-path2d';
-import { getTheme } from './theme';
 import { ResourceLoader } from '../resource-loader/loader';
 import { AttributeUpdateType, IContainPointMode, UpdateTag } from '../common/enums';
+import { BoundsContext } from '../common/bounds-context';
+import { renderCommandList } from '../common/render-command-list';
+import { parsePadding } from '../common/utils';
 
 /**
  * pathProxy参考自zrender
@@ -213,8 +215,8 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
   declare in3dMode?: boolean;
 
   // aabbBounds，所有图形都需要有，所以初始化即赋值
-  protected declare _AABBBounds: AABBBounds;
-  get AABBBounds(): AABBBounds {
+  protected declare _AABBBounds: IAABBBounds;
+  get AABBBounds(): IAABBBounds {
     return this.tryUpdateAABBBounds(this.attribute.boundsMode === 'imprecise');
   }
   // 具有旋转的包围盒，部分图元需要，动态初始化
@@ -222,9 +224,9 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
   // get OBBBounds(): OBBBounds {
   //   return this.tryUpdateOBBBounds();
   // }
-  protected declare _globalAABBBounds: AABBBounds;
+  protected declare _globalAABBBounds: IAABBBounds;
   // 全局包围盒，部分图元需要，动态初始化，建议使用AABBBounds
-  get globalAABBBounds(): AABBBounds {
+  get globalAABBBounds(): IAABBBounds {
     return this.tryUpdateGlobalAABBBounds();
   }
   protected declare _transMatrix: Matrix;
@@ -307,7 +309,7 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
     this._emitCustomEvent('animate-bind', animate);
   }
 
-  protected tryUpdateAABBBounds(full?: boolean): AABBBounds {
+  protected tryUpdateAABBBounds(full?: boolean): IAABBBounds {
     if (!this.shouldUpdateAABBBounds()) {
       return this._AABBBounds;
     }
@@ -331,9 +333,42 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
     }
   }
 
-  protected abstract doUpdateAABBBounds(full?: boolean): AABBBounds;
+  abstract getGraphicTheme(): T;
 
-  protected tryUpdateGlobalAABBBounds(): AABBBounds {
+  protected abstract updateAABBBounds(
+    attribute: T,
+    symbolTheme: Required<T>,
+    aabbBounds: IAABBBounds,
+    full?: boolean
+  ): IAABBBounds;
+
+  protected doUpdateAABBBounds(full?: boolean): IAABBBounds {
+    const graphicTheme = this.getGraphicTheme();
+    this._AABBBounds.clear();
+    const attribute = this.attribute;
+    const bounds = this.updateAABBBounds(attribute, graphicTheme as Required<T>, this._AABBBounds) as AABBBounds;
+
+    const { boundsPadding = graphicTheme.boundsPadding } = attribute;
+    const paddingArray = parsePadding(boundsPadding);
+    if (paddingArray) {
+      bounds.expand(paddingArray);
+    }
+
+    this.clearUpdateBoundTag();
+    return bounds;
+  }
+
+  updatePathProxyAABBBounds(aabbBounds: IAABBBounds): boolean {
+    const path = typeof this.pathProxy === 'function' ? (this.pathProxy as any)(this.attribute) : this.pathProxy;
+    if (!path) {
+      return false;
+    }
+    const boundsContext = new BoundsContext(aabbBounds);
+    renderCommandList(path.commandList, boundsContext, 0, 0);
+    return true;
+  }
+
+  protected tryUpdateGlobalAABBBounds(): IAABBBounds {
     const b = this.AABBBounds;
     if (!this._globalAABBBounds) {
       this._globalAABBBounds = b.clone();
@@ -463,7 +498,7 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
     for (let i = 0; i < keys.length; i++) {
       const key = keys[i];
 
-      this.attribute[key] = params[key];
+      (this.attribute as any)[key] = (params as any)[key];
     }
     this.valid = this.isValid();
     // 没有设置shape&bounds的tag
@@ -481,10 +516,10 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
     const params =
       this.onBeforeAttributeUpdate && this.onBeforeAttributeUpdate({ [key]: value }, this.attribute, key, context);
     if (!params) {
-      if (!isNil(this.normalAttrs?.[key])) {
-        this.normalAttrs[key] = value;
+      if (!isNil((this.normalAttrs as any)?.[key])) {
+        (this.normalAttrs as any)[key] = value;
       } else {
-        this.attribute[key] = value;
+        (this.attribute as any)[key] = value;
         this.valid = this.isValid();
         if (!this.updateShapeAndBoundsTagSetted() && (forceUpdateTag || this.needUpdateTag(key))) {
           this.addUpdateShapeAndBoundsTag();
@@ -785,12 +820,13 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
 
       keys.forEach(key => {
         if (!noWorkAttrs[key]) {
-          animateAttrs[key] = isClear && attrs[key] === undefined ? this.getDefaultAttribute(key) : attrs[key];
+          (animateAttrs as any)[key] =
+            isClear && (attrs as any)[key] === undefined ? this.getDefaultAttribute(key) : (attrs as any)[key];
         } else {
           if (!noAnimateAttrs) {
             noAnimateAttrs = {};
           }
-          noAnimateAttrs[key] = attrs[key];
+          (noAnimateAttrs as any)[key] = (attrs as any)[key];
         }
       });
 
@@ -813,18 +849,18 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
     if (this.normalAttrs) {
       Object.keys(stateAttrs).forEach(key => {
         if (key in this.normalAttrs) {
-          newNormalAttrs[key] = this.normalAttrs[key];
-          delete this.normalAttrs[key];
+          (newNormalAttrs as any)[key] = (this.normalAttrs as any)[key];
+          delete (this.normalAttrs as any)[key];
         } else {
-          newNormalAttrs[key] = this.getNormalAttribute(key);
+          (newNormalAttrs as any)[key] = this.getNormalAttribute(key);
         }
       });
       Object.keys(this.normalAttrs).forEach(key => {
-        stateAttrs[key] = this.normalAttrs[key];
+        (stateAttrs as any)[key] = (this.normalAttrs as any)[key];
       });
     } else {
       Object.keys(stateAttrs).forEach(key => {
-        newNormalAttrs[key] = this.getNormalAttribute(key);
+        (newNormalAttrs as any)[key] = this.getNormalAttribute(key);
       });
     }
 
@@ -843,7 +879,7 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
   }
 
   private getNormalAttribute(key: string) {
-    let value = this.attribute[key];
+    let value = (this.attribute as any)[key];
 
     if (this.animates) {
       this.animates.forEach(animate => {
@@ -1295,12 +1331,12 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
   }
 
   getDefaultAttribute(name: string) {
-    return getTheme(this as any)[this.type][name];
+    return (this.getGraphicTheme() as any)[name];
   }
 
   // 获取属性
   getComputedAttribute(name: string) {
-    return this.attribute[name] ?? this.getDefaultAttribute(name);
+    return (this.attribute as any)[name] ?? this.getDefaultAttribute(name);
   }
   /**
    * 添加onSetStage钩子
