+++ packages/vrender-kits/src/render/contributions/rough/rough-arc.ts
@@ -50,7 +50,7 @@ export class RoughCanvasArcRender extends RoughBaseRender implements IGraphicRen
     context.highPerformanceSave();
 
     // const arcAttribute = graphicService.themeService.getCurrentTheme().arcAttribute;
-    const arcAttribute = getTheme(arc).arc;
+    const arcAttribute = arc.getGraphicTheme();
     let { x = arcAttribute.x, y = arcAttribute.y } = arc.attribute;
     if (!arc.transMatrix.onlyTranslate()) {
       // 性能较差
